+++ Adapters/BidonAdapterMintegral/Providers/Bidding/MintegralBiddingAdViewDemandProvider.swift
@@ -100,7 +100,7 @@ extension MintegralBiddingAdViewDemandProvider: MTGBannerAdViewDelegate {
     }
 
     func adViewClosed(_ adView: MTGBannerAdView!) {
-        delegate?.providerDidHide(self)
+        // No-op: Banner ads don't use providerDidHide - that's for fullscreen ads only
     }
 }